Default to always (even privileged) run in a user namespace
Since we're setting NO_NEW_PRIVS, and are not generally running
bubblewrap as the root user there is very little use for other uids.

Additionally, this is the only mode which works when using
unprivileged user namespaces, so it makes sense to use it by default.

You can use --share-user to not use a user namespace, but that will
not work unless bubblewrap is setuid.