Add configure flags to allow setting setuid/setcaps
With this you can e.g. :
./configure --enable-sudo --with-priv-mode=setcaps
make
make install

and it will ask you for sudo password and then make the final binary
have the right capabilities set.

This is not needed when setting such persmissions in e.g. a spec file, but
it is useful for developers building bubblewrap.

Pull request: #26
Approved by: cgwalters