Swap --share-user with --unshare-user, but auto-unshare if !setuid
We have to support two different ways to run:
 - As setuid root, for systems without unprivileged userns support
 - Non-setuid, but require unprivileged userns

The fact that we exposed `--share-user` is awkward, because it forced
tools that want to work in both case to basically reimplement the
logic for detecting userns support, if they didn't care whether
or not userns was enabled.

For example in the case of `demos/bubblewrap-shell.sh` where we
share the invoking UID.

This commit changes things so we now default to `--unshare-user` if
we're *not* installed privileged, since it's a requirement.

The end result here is that we just work out of the box in more
scenarios; callers that require the uid mapping portion of userns will
still be passing `--uid`, and this will still properly fail if the
kernel doesn't have userns.

Closes: #36