Propagate exit status when not in a PID namespace
If we're not doing a PID namespace, we don't create a monitor
process, which means that the code in `monitor_child()` needs
to properly propagate the exit status from the signalfd.

It might be better to change `monitor_child()` to be a `waitpid()`
loop in this case, but I decided to go for the one liner fix that's an
improvement in both cases anyways.

I noticed this with:

```
bwrap --ro-bind / / --dev /dev true
```

exiting with code 1.

Closes: #49
Approved by: rhatdan