README: update references to runC
Currently we are working on implementing rootless containers, where no
privileges are required during any part of the process of installing
runC or the management of containers. We are solving a different problem
to bubblewrap with this feature, in the hopes that users on machines
where they have no ability to create a setuid binary will be able to
still use containers.

Signed-off-by: Aleksa Sarai <asarai@suse.de>

Closes: #84
Approved by: cgwalters