Switch to using uid 0 + fsuid == uid instead of capabilities
We used to drop back to the real user immediately, keeping only
a few capabilities, which did *not* include CAP_DAC_OVERRIDE, so that
any code in the privileged parts would be unable to read or write
to files that the user otherwise couldn't.

However, running the privileged code as the user turns out to be
problematic in the case of --unshare-user, as it gives *any* process
in the parent user namespace (i.e. the regular one) privileges to
ptrace all processes in the user namespace, including the privileged
setup code. This means that one could run bwrap --unshare-user and then
ptrace the child and do anything with full privileges in the child
user namespace.

This is not necessarily a huge problem, as the kernel shouldn't allow
you to do anything you couldn't outside the user namespace. However,
not everyone trusts the user namespace implementation, and one of the
points of bubblewrap is to be able to securely allow a small subset
of user namespaces, and this was voided by this ptrace ability.

So, instead we run the entire setup as euid 0, which means that
nothing (even in a user ns) is allowed to ptrace it (because euid !=
uid, and we're not dumpable). To protect against the user accessing
files not accessible to him we set fsuid to the real uid.