Handle inherited children dying
Its quite possible for bwrap to have children to it exit and deliver
SIGCHLD signals and the like. For instance, the bubblewrap-shell
demo does (simplified):

(exec bwrap --file 11 /a/file /bin/sh) 11< <(cmd)

This actually starts a subshell process, and has it spawn a child
running "cmd", and then exec replaces the subshell with the bwrap
process that now has an unexpected child.

This can lead to problem in two ways:

 1) We get a SIGCHLD before we process children dying via
    the signalfd. This means we never wait() for the child
    and it becomes a zombie.

 2) If the child dies after the pid1 process has started
    we will see the sigchld and think that pid1 has died
    and exit the monitor process.

The fix for 1 is to reap all outstanding zombies after we
blocked sigchld, but before we fork the child.

The fix for 2 is to only exit when the expected pid dies.

Closes: #146
Approved by: cgwalters