Clear capability bounding set
The capability bounding set is a limit on what capabilities can
be regained at execve(). Due to PR_NO_NEW_PRIVS we should be safe
from any such issues, but we may as well clear it anyway.

Note, we also have to clear it in the new namespace if user namespaces
are enabled, because the kernel gives us a new set of full bounds in
the user namespace.

See https://github.com/projectatomic/bubblewrap/issues/136 for some
discussion about this.

Closes: #149
Approved by: cgwalters