Use seccomp instead of setsid() to workaround CVE-2017-5226
The setsid() workaround of
https://github.com/projectatomic/bubblewrap/pull/143 is problematic,
because it e.g. breaks shell job control for bubblewrap instances.
So, instead we use a seccomp approach based on:
 https://github.com/karelzak/util-linux/commit/8e4925016875c6a4f2ab4f833ba66f0fc57396a2
However, since we don't want to pull in any more dependencies into
the setuid binary we pre-compile the seccomp code during the build.

If libseccomp is not available on your architecture, we still support
the old fix with --disable-seccomp-tty-fix.

This fixes https://github.com/projectatomic/bubblewrap/issues/147