tests: Add a test case for recursive bwrap
Now that we have `--cap-add ALL`, and the previous commit fixed
things to use `pivot_root`, we can cleanly support recursive `bwrap`
invocations.  This is useful for a variety of cases; today for example
we often run `rpm-ostree compose` inside docker, but there's no reason
the outer container couldn't be bwrap based.