Make --bind-data files have permissions 0666
When you use --bind-data or --ro-bind-data the file currently ends up with
permission 0600, as this is the mkstemp() default. All other codepaths in
bubblewrap creates 0666 files so that other users can read the files,
which is more common than top-secret files only readable by the user.

We fix this by fchmod()ing to 0666