Add support for --socket FD PATH
This bind()s and listen()s to a AF_UNIX socket created by socket() at
a specified location in the sandbox. This is useful if you want to
have a socket available in the sandbox talking to a service outside
the sandbox without ever exposing the socket path outside the sandbox.

In particular, flatpak wants to pre-create sockets for all dbus proxies
and have them bound *only* in the sandbox.