Add --close-fd argument
This just closes the fd once the sandbox is setup, and lets you know
that e.g. any sockets are bound and listened.

Flatpak will use this to ensure that all dbus proxy sockets are
bound and listened before spawning the proxy (because its an error
to call accept() on the sockets before its been listen()ed).