Fix leak detected by LSan/ASan
Some variables like base_path ("/run/user/%d/.bubblewrap") are
declared with the cleanup attribute in main(), but this cleanup is not
run when in the parent process, since it calls exit() in monitor_child().

Use return statements instead of exit() so that cleanup attributes
will be run.

Closes: #271
Approved by: smcv