WSL: Disable read-only bind mount
When re-mounting a bind mount read-only WSL fails with EBUSY if there
are any outstanding writable file-descriptors anywhere on the filesystem
(even outside the bind-mount). This makes read-only bindmounts fail
most of the time (for example in flatpak due to writable /var/tmp files).

For now we just disable read-only.