ci: Bump to fedora/29/atomic
Fedora 27 has been EOL for a while now... bump it to 29.

It'd be nice to test against FCOS, though it's not currently supported
by PAPR. It's in maintenance mode, but I might add support for it as a
stopgap. Medium-term though, I want to deprecate PAPR in favour of other
CI solutions.

Anyway, for now hopefully this should fix the CI on this repo.

Closes: #327
Approved by: alexlarsson