setuid mode: Properly drop privs in monitor and pid1
It turns out we have this check in drop_privs():

 if (getuid () == 0 && setuid (opt_sandbox_uid) < 0)

Which is supposed to drop back to the regular uid in the case
we're in setuid mode and we're in the monitor_child() or do_init()
processes.

Unfortunately we're setuid, not plain root, so uid is not 0, but euid is zero.

This caused the monitoring processes to be running partially as root
which shows up weird in /proc.

Fix this by checking euid for 0 instead.