introduce live mounting
In flatpak we would like to be able to make a host subtree visible to
some bwrap'ed processes some time after they have already been started
(https://github.com/flatpak/xdg-desktop-portal/issues/200).

This patch introduces the '--live-mount-fd FD' option. The process
calling bwrap can pass a FD this way and write pairs of directories to
it which will be interpreted as source and destionation of a bind mount
operation.

To make this work in bwrap we make /newroot a shared mount and clone off
into a new mount namespace for what eventually becomes the container,
make /newroot a slave mount in this new mount namespace and continue on
normally with unmounting /oldroot and pivioting to /newroot. The original
process now still has /oldroot and /newroot and changes in /newroot
propagate to the container's mount namespace.