Add --foreground option
This option creates a new process group for the child process and
if there's a controlling terminal, makes the new process group the
foreground process group of the controlling terminal. This makes
sure that SIGINT signals generated by CTRL+C are only sent to the
child process and not to bubblewrap itself.

While this can also be achieved with external utilities (e.g.
https://github.com/util-linux/util-linux/pull/2445), it still makes
sense to support this natively in bubblewrap to avoid the need to
install any other utilities in the sandbox to be able to make the
child process the foreground process group of the controlling
terminal.

Fixes #369

Signed-off-by: Daan De Meyer <daan.j.demeyer@gmail.com>